Reorganized Drone Control to Allow for Multiple Drones
Created PlayerController, which will send input to the correct drone and spawn Drones.
Created SpawnArea script, which defines a trigger that detects whether an item can be spawned in that area. Created a SpawnArea Prefab which is a generic game object with the BoxCollider and Rigidbody configured to correctly detect intersecting bodies.
Moved some assets into more specific folders